Dockerfile_build: supply configs and environment variables
Don't try to build with one libc and run with another.

Don't disable cgo - it's a hard requirement for setting up user
namespaces.  The default seccomp filter for dockerd doesn't allow us to
use unshare() as part of setting up namespaces, so if we're running
under dockerd, we still need to be started with either
	--security-opt seccomp=unconfined
to not use a seccomp filter, or
	--security-opt seccomp=/usr/share/containers/seccomp.json
to use the seccomp filter that podman uses, which will allow the
container to use the unshare() syscall.

Supply additional configuration files expected by libraries.

Signed-off-by: Nalin Dahyabhai <nalin@redhat.com>